Split run_cluster.py into cluster_evolution.py and run_cluster.py
